Implement a self-contained vector/matrix operations API
We already had some of this inside the cursor implementation, this pulls it out,
and makes it more general and complete, so you can use it by itself. Main use
case is manipulating block structures.

This also adds `xyz` and `with-xyz`, making it easier to manipulate
location-like things in a uniform way.