Make our traits stateful, and honor load/save
This provides a mechanism for Clojure-based trait implementations to carry state
by making the Trait instance quack like an Atom. It implements default load/save
methods that handle persisting this state, so that this is handled
transparently (see `plugins/Citizens/saves.yml`).

This introduces two new values for the options map to `make-trait`: `:init` and
`:post-init`. `:init` is a map or zero-arg function returning a map, which is
used to set the initial state when instantiating a new trait. `:post-init`
corresponds with `gen-class`'s `:post-init`, it's a single-arg function that
receives the instantiated Trait object.