Implement Adventure text-component based display name handling
PaperMC now integrates Adventure for these things, and considers the string
based methods deprecated. This hasn't been much of an issue, but it can cause
issues because the way Paper converts a plain string to a component contains an
extra wrapping component, and that breaks equality checks on item stacks.

We'll have to do this properly at some point, implement two-way markup/component
conversion, and handle both getting and setting of lore and displayname. For now
doing the 80% thing: correctly setting the display name.